Added zellij config and layout
Added a zellij config file and a default layout for development. The
installation of both were added to the install script. The default
layout is quite clunky at the moment due to a bug in zellij 0.40.1 which
I have opened an issue for
(https://github.com/zellij-org/zellij/issues/3632). The layout should be
changed to properly use `default_tab_template` when the issue is
resolved.