Replace gen.nvim with gp.nvim
Changed the plugin used to interact with local LLM models. The gp.nvim
plugin is more plug and play and has out of the box support for letting
models edit buffers directly. This plugin does not provide completion,
but at the moment I feel like that feature is too invasive anyways and
prefer to command models when I please.