Removed gen.nvim AI plugin
Removed the gen.nvim plugin responsible for interfacing with an ollama
model. This capability was moved to a separate branch so that the main
branch can be used to set up a dev environment on machines that are not
capable of running models locally. The ai-code-gen branch will be kept
up-to-date with the main branch but include changes to enable LLM
integration.