Created environment scripts under .bashrc.d/
Moved the existing neovim environment script under .bashrc.d/. Set up a
zellij environment script that allows for adding projects to a list that
can be used to quickly create sessions with a known ID, at a path
location, and with a layout. If a project cannot be found, zellij is run
like normal.