Rework zellij session management
Resurrecting zellij sessions can sometimes lead to some odd behavior
that is not worth dealing with when it is just as easy (in this use
case) to create a new session. However, detaching and reattaching is
very nice so I added a keybind to detach rather than quit entirely and
allowed the project picker to attach to detached sessions.