Created custom zellij status using zjstatus plugin
Created a zellij status bar with custom styling for the tabs and extra
widgets in the top right for session name, battery, and time. Removed
the keybind helper bar from the bottom of the default layout. Created a
script to produce the battery capacity for the status bar. Modified the
installation script to download the zjstatus plugin and copy it and any
widget scripts along with the zellij config.